Add files via upload
Assignment Description:

In this SQL Assignment, we will be working with a simulated dataset inspired by the operations of a ride-hailing service like Uber. The dataset consists of information related to rides, drivers, and passengers. Our goal is to analyze and gain insights from this data using SQL queries and techniques.